requirements.txt
certifi==2021.10.8
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.9
django-cors-headers==2.4.0
django-rest-swagger==2.1.0
djangorestframework==3.5.1
idna==2.10
inflection==0.3.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
pyparsing==2.4.7
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
ruamel.ordereddict==0.4.15
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.6
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.8
